+++ packages/uno-serverless-cli/test/unit/samples/entities-path/address.ts
@@ -1,0 +1,8 @@
+import { Xyz } from "@entities/xyz";
+
+export interface Address {
+  street: string;
+  city: string;
+  zip: number;
+  xyz: Xyz;
+}