+++ packages/uno-serverless-cli/test/unit/samples/entities-path/user.ts
@@ -1,0 +1,20 @@
+import { Address } from "@entities/address";
+
+export interface User {
+  address?: Address;
+  /**
+   * @format email
+   */
+  email: string;
+  /**
+   * @minLength 6
+   */
+  password: string;
+  firstName?: string;
+  lastName?: string;
+  composite?: Array<{
+    address: Address;
+    id: string,
+    name?: string;
+  }>;
+}